requirements.txt
asgiref
certifi
cffi
celery
charset-normalizer
cryptography
datetime-truncate
defusedxml
Deprecated
Django
boto3
django-storages 
django-admin-chart
django_redis
django-jazzmin
django-admin-interface
django-bower
django-braces
django-chartjs
django-colorfield
django-cors-headers
django-memoize
django-multiselectfield
django-oauth-toolkit
djangorestframework
djangorestframework-simplejwt
recently_visited_page_middleware
drf-yasg
fcm-django
gunicorn
geoip2
idna
inflection
Jinja2
jwcrypto
MarkupSafe
oauthlib
packaging
Pillow
psycopg2-binary
pycparser
PyJWT
python-dateutil
python-decouple
python-slugify
python3-openid
pytz
PyYAML
pyotp
requests
requests-oauthlib
redis
six
sqlparse
text-unidecode
typing_extensions
Unidecode
uritemplate
urllib3
whitenoise
wrapt
braintree

